move operator code out of __init__ to avoid loading unused code in model server startup
